Only install /etc/init.d/rc.d/* if systemd-sysv-install is not present

+++ rpms/SPECS/wazuh-agent.spec
@@ -81,14 +81,16 @@ echo 'USER_AUTO_START="n"' >> ./etc/preloaded-vars.conf
 %endif
 
 # Create directories
-mkdir -p ${RPM_BUILD_ROOT}%{_initrddir}
+if [ ! -f /usr/lib/systemd/systemd-sysv-install ]; then mkdir -p ${RPM_BUILD_ROOT}%{_initrddir}; fi
 mkdir -p ${RPM_BUILD_ROOT}%{_localstatedir}/.ssh
 
 # Copy the installed files into RPM_BUILD_ROOT directory
 cp -pr %{_localstatedir}/* ${RPM_BUILD_ROOT}%{_localstatedir}/
 mkdir -p ${RPM_BUILD_ROOT}/usr/lib/systemd/system/
-sed -i "s:WAZUH_HOME_TMP:%{_localstatedir}:g" src/init/templates/ossec-hids-rh.init
-install -m 0755 src/init/templates/ossec-hids-rh.init ${RPM_BUILD_ROOT}%{_initrddir}/wazuh-agent
+if [ ! -f /usr/lib/systemd/systemd-sysv-install ]; then
+  sed -i "s:WAZUH_HOME_TMP:%{_localstatedir}:g" src/init/templates/ossec-hids-rh.init
+  install -m 0755 src/init/templates/ossec-hids-rh.init ${RPM_BUILD_ROOT}%{_initrddir}/wazuh-agent
+fi
 sed -i "s:WAZUH_HOME_TMP:%{_localstatedir}:g" src/init/templates/wazuh-agent.service
 install -m 0644 src/init/templates/wazuh-agent.service ${RPM_BUILD_ROOT}/usr/lib/systemd/system/
 
@@ -240,12 +242,6 @@ if [ $1 = 1 ]; then
   %{_localstatedir}/packages_files/agent_installation_scripts/src/init/register_configure_agent.sh %{_localstatedir} > /dev/null || :
 fi
 
-if [ -f /etc/os-release ]; then
-  source /etc/os-release
-  if [ "${NAME}" = "Red Hat Enterprise Linux" ] && [ "$((${VERSION_ID:0:1}))" -ge 9 ]; then
-    rm -f %{_initrddir}/wazuh-agent
-  fi
-fi
 
   # We create this fix for the operating system that deprecated the SySV. For now, this fix is for suse/openSUSE
   sles=""
